+++ RELEASE_NOTES.md
@@ -1,6 +1,6 @@
-
 <!-- Keep them in alphabetical order -->
 ## Enhancements
+- [KOGITO-8452](https://issues.redhat.com/browse/KOGITO-8452) - [SW] Update Serverless Operator to use CNCF go-sdk model
 
 ## Bug Fixes